HallEffectSensor and FingerTracking library work with prototype glove
